conveted some floats to decimals

+++ nadine/forms.py
@@ -603,7 +603,7 @@ class RoomForm(forms.Form):
     max_capacity = forms.IntegerField(min_value=1, max_value=1000, required=True)
     has_av = forms.BooleanField(required=False)
     has_phone = forms.BooleanField(required=False)
-    default_rate = forms.FloatField(required=True, min_value=0, max_value=None)
+    default_rate = forms.DecimalField(required=True, min_value=0, max_value=None)
     image = forms.FileField(required=False)
     members_only = forms.BooleanField(required=False)
 
@@ -719,4 +719,3 @@ class MembershipPackageForm(forms.Form):
         return sub_default
 
 # Copyright 2021 Office Nomads LLC (https://officenomads.com/) Licensed under the AGPL License, Version 3.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at https://www.gnu.org/licenses/agpl-3.0.html. Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
